Create separate component for followed options
